moved categoryList to category component folder
