+++ src/views/ApplicationViews.js
@@ -2,7 +2,7 @@ import { Route, Routes } from "react-router-dom"
 import { Login } from "../components/auth/Login"
 import { Register } from "../components/auth/Register"
 import { Authorized } from "./Authorized"
-import { CategoryList } from "../components/categories/CategoryList"
+import { CategoryList } from "../components/category/CategoryList"
 import { CreatePost } from "../components/posts/CreatePost"
 import CreateCategory from "../components/category/createCategory"
